NetworkWriter.Position type changed from 'short' to 'int' to fix a bug where writing >32kb into NetworkWriter would result in negative .Position

+++ Unity-Technologies-networking/Runtime/NetworkIdentity.cs
@@ -741,7 +741,7 @@ namespace UnityEngine.Networking
                     s_UpdateWriter.Write(netId);
 
                     bool wroteData = false;
-                    short oldPos;
+                    int oldPos;
                     for (int i = 0; i < m_NetworkBehaviours.Length; i++)
                     {
                         oldPos = s_UpdateWriter.Position;
